verbs: Move message interface into separate file
